Update README.md
The SLAM (Simultaneous Localization and Mapping) is a technique to draw a map by estimating current location in an arbitrary space. The SLAM is a well-known feature of TurtleBot from its predecessors. The video here shows you how accurately TurtleBot3 can draw a map with its compact and affordable platform.

All SLAM solutions include some kind of device or tool that allows a robot or other vehicle to observe and measure the environment around it.

This can be done by cameras, other types of image sensors, LiDAR laser scanner technology and even sonar. Essentially, any device that can be used to measure physical properties like location, distance or velocity can be included as part of a SLAM system. 